+++ src/processing/filters/mod.rs
@@ -1,23 +1,23 @@
-#[cfg(feature = "ai")]
-pub mod ai;
 pub mod counter;
 pub mod duration;
 pub mod file_size;
 pub mod incoming;
 pub mod message_type;
+#[cfg(feature = "openai")]
+pub mod openai;
 pub mod regexp;
 pub mod text_length;
 #[cfg(feature = "storage")]
 pub mod unique;
 pub mod wordlist;
 
-#[cfg(feature = "ai")]
-pub(crate) use ai::Context;
 pub(crate) use counter::Counter;
 pub(crate) use duration::Duration;
 pub(crate) use file_size::FileSize;
 pub(crate) use incoming::Incoming;
 pub(crate) use message_type::MessageType;
+#[cfg(feature = "openai")]
+pub(crate) use openai::OpenAi;
 pub(crate) use regexp::Regexp;
 pub(crate) use text_length::TextLength;
 #[cfg(feature = "storage")]