+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 ===
 ## v0.7.0
-* new filter - `Context` filter with `openai` models
+* new filter - `OpenAi`
 ## v0.6.0
 * new pipe - `Format`
 ## v0.5.0